Enroll *.img files in license protection
Tests with the new Django codebase have shown that u-boot binaries would
not be protected by the license wrapper. While this is not a problem for
u-boot it may become a problem for fastboot images (for example built
for the galaxy nexus device).

To prevent tis problem I've added a new item to the 'File-Pattern' list
to match all '*.img' files.

Acked-by: Georgy Redkozubov <georgy.redkozubov@linaro.org>